+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGES
 
+## Unreleased
+
+* Bevy `0.7`
+* `Sprite` supports `bevy_sprite::Anchor`
+* `TextureAtlasSprite` support
+* using `copyless` crate
+
 ## 0.1.0
 
 * First version